requirements.txt
annotated-types==0.7.0
certifi==2025.6.15
cycler==0.12.1
distro==1.9.0
filelock==3.18.0
fsspec==2025.5.1
httmock==1.4.0
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
prometheus-api-client==0.6.0
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
safetensors==0.5.3
scipy==1.15.3
setuptools==80.9.0
sniffio==1.3.1
threadpoolctl==3.6.0
torch==2.7.1
transformers==4.52.4
typing-inspection==0.4.1
tzlocal==5.3.1
sentence-transformers==5.1.1
scikit-learn==1.7.2
Jinja2==3.1.6
python-dotenv==1.1.1
numpy==2.3.3